requirements.txt
dnspython==1.16.0
lxml==4.5.0
mysql-connector-python==8.0.19
protobuf==3.6.1
python-docx==0.8.10
six==1.14.0
